requirements.txt
numpy
Pillow
pandas
opencv-python
torch
torchvision
pytorch-lightning
torchmetrics
openpyxl